+++ src/main/java/team05/integrated_feed_backend/module/auth/entity/VerificationCode.java
@@ -3,7 +3,7 @@ package team05.integrated_feed_backend.module.auth.entity;
 import jakarta.persistence.*;
 import lombok.*;
 import team05.integrated_feed_backend.common.entity.BaseEntity;
-import team05.integrated_feed_backend.module.user.entity.Member;
+import team05.integrated_feed_backend.module.member.entity.Member;
 
 import java.time.LocalDateTime;
